fix: clean architecture and finish discord and telegram implementation
